Revert "Minor refactors"
