--- .github/workflows/lint.yml
@@ -1,26 +1,0 @@
-# This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
-name: Lint
-
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Setup Node.js 10.x
-        uses: actions/setup-node@v1
-        with:
-          node-version: 10
-      - name: Install packages
-        run: yarn install
-      - name: Run ESLint
-        run: yarn lint-no-fix
-
-
