+++ .github/workflows/main.yml
@@ -1,6 +1,7 @@
 # This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
-name: Tests
+
+name: Node.js CI
 
 on:
   push:
@@ -10,26 +11,23 @@ on:
     branches: [ master ]
 
 jobs:
-  jest:
+  test:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
-    - name: Setup Node.js 10.x
+    - name: Use Node.js 10
       uses: actions/setup-node@v1
       with:
-        node-version: 10
-    - name: Install packages
+        node-version: '12'
+    - name: Install dependencies
       run: yarn install
-    - name: Run tests
-      run: yarn test
-      env:
-        DISABLE_TRANSACTIONS: TRUE
-    - name: Upload code coverage to Code Climate
+    - name: test with coverage
       uses: paambaati/codeclimate-action@v2.6.0
       env:
         CC_TEST_REPORTER_ID: ${{ secrets.CC_TEST_REPORTER_ID }}
+        DISABLE_TRANSACTIONS: TRUE
       with:
-        coverageCommand: "echo -" # A useless command since the coverage is already generated in the previous steps.
+        coverageCommand: yarn test
         coverageLocations: ${{github.workspace}}/coverage/lcov.info:lcov
