+++ README.md
@@ -1,8 +1,8 @@
 # Somalia Code
 
-![Tests](https://github.com/flatten-official/flatten-somalia/workflows/Tests/badge.svg)
-[![Maintainability](https://api.codeclimate.com/v1/badges/a195459a45a7e562ac07/maintainability)](https://codeclimate.com/repos/5eebb93769ce914dc100dcd1/maintainability)
-[![Test Coverage](https://api.codeclimate.com/v1/badges/a195459a45a7e562ac07/test_coverage)](https://codeclimate.com/repos/5eebb93769ce914dc100dcd1/test_coverage)
+![backend-tests](https://github.com/flatten-official/flatten-somalia/workflows/Node.js%20CI/badge.svg)
+<a href="https://codeclimate.com/repos/5eebb93769ce914dc100dcd1/maintainability"><img src="https://api.codeclimate.com/v1/badges/a195459a45a7e562ac07/maintainability" /></a>
+<a href="https://codeclimate.com/repos/5eebb93769ce914dc100dcd1/test_coverage"><img src="https://api.codeclimate.com/v1/badges/a195459a45a7e562ac07/test_coverage" /></a>
 
 This repository contains the code used in Somalia.
