+++ package.json
@@ -8,11 +8,10 @@
   "scripts": {
     "dev-backend": "yarn workspace backend dev",
     "dev-frontend": "node scripts/frontend/start.js",
-    "dev": "concurrently -k -i \"yarn dev-backend\" \"yarn dev-frontend\"",
+    "dev": "concurrently -k \"yarn dev-backend\" \"yarn dev-frontend\"",
     "build-frontend": "node scripts/frontend/build.js",
-    "serve-frontend": "serve --single --listen 80 ./build",
+    "serve-frontend": "serve --single --listen 3000 ./build",
     "lint": "eslint --fix .",
-    "lint-no-fix": "eslint .",
     "test-cloud-func": "node scripts/cloudFunctionLauncher.js",
     "auth": "gcloud auth login && gcloud auth application-default login && gcloud config set project flatten-staging-271921",
     "test": "jest",
@@ -57,7 +56,6 @@
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-react": "^7.20.6",
     "eslint-plugin-react-hooks": "^4.1.0",
-    "eslint-plugin-workspaces": "^0.5.5",
     "file-loader": "^6.0.0",
     "fs-extra": "^9.0.1",
     "html-webpack-plugin": "^4.3.0",
