+++ packages/backend/src/index.js
@@ -24,7 +24,7 @@ async function setup() {
 }
 
 async function startServer() {
-  const port = process.env.BACKEND_PORT || 3000;
+  const port = process.env.PORT || 80;
 
   const app = await getApp();
   app.listen(port, () => {
