+++ packages/frontend/package.json
@@ -30,7 +30,6 @@
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "seamless-immutable": "^7.1.4",
-    "util-shared-constants": "1.0.0"
+    "seamless-immutable": "^7.1.4"
   }
 }
