+++ packages/frontend/src/api/backend.dev.js
@@ -1,1 +1,1 @@
-export default "http://localhost:3000/";
+export default "http://localhost/";
