+++ packages/frontend/src/api/constants.js
@@ -1,0 +1,5 @@
+export const Permissions = {
+  submitForms: "submitForms",
+  manageVolunteers: "manageVolunteers",
+  access: "access",
+};
