+++ packages/frontend/src/api/endpoints.js
@@ -1,6 +1,4 @@
 import axios from "axios";
-// (incorrectly flagged because of name collision)
-// eslint-disable-next-line workspaces/require-dependency
 import backend from "./backend";
 
 const axiosClient = axios.create({ baseURL: backend, withCredentials: true });
