+++ packages/frontend/src/ui/commonComponents/formio/Form.js
@@ -38,8 +38,6 @@ const Form = ({
       await submitHook(submission.data);
       setSubmitted(true);
     } catch (e) {
-      // suppress warning as it can be useful even in production to debug such errors
-      // eslint-disable-next-line no-console
       console.error(e);
       next(e);
     }
