--- packages/util-shared-constants/README.md
@@ -1,3 +1,0 @@
-# util-shared-constants
-
-This package contains constants that are shared between the frontend and the backend (e.g. permission strings).
