--- packages/util-shared-constants/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "util-shared-constants",
-  "version": "1.0.0",
-  "main": "index.js",
-  "private": true
-}
